Define DEFAULT_CONFIG only once, move to yolo/
