+++ ultralytics/yolo/configs/__init__.py
@@ -3,7 +3,7 @@ from typing import Dict, Union
 
 from omegaconf import DictConfig, OmegaConf
 
-from ultralytics.yolo.utils.configs.hydra_patch import check_config_mismatch
+from ultralytics.yolo.configs.hydra_patch import check_config_mismatch
 
 
 def get_config(config: Union[str, DictConfig], overrides: Union[str, Dict] = {}):
