+++ ultralytics/yolo/engine/predictor.py
@@ -32,14 +32,12 @@ import cv2
 
 from ultralytics.nn.autobackend import AutoBackend
 from ultralytics.yolo.data.dataloaders.stream_loaders import LoadImages, LoadScreenshots, LoadStreams
-from ultralytics.yolo.data.utils import IMG_FORMATS, VID_FORMATS, check_dataset, check_dataset_yaml
-from ultralytics.yolo.utils import LOGGER, ROOT, colorstr, ops
+from ultralytics.yolo.data.utils import IMG_FORMATS, VID_FORMATS
+from ultralytics.yolo.utils import DEFAULT_CONFIG, LOGGER, colorstr, ops
 from ultralytics.yolo.utils.checks import check_file, check_imshow
-from ultralytics.yolo.utils.configs import get_config
+from ultralytics.yolo.configs import get_config
 from ultralytics.yolo.utils.files import increment_path
 from ultralytics.yolo.utils.torch_utils import check_imgsz, select_device, smart_inference_mode
-
-DEFAULT_CONFIG = ROOT / "yolo/utils/configs/default.yaml"
 
 
 class BasePredictor:
