+++ ultralytics/yolo/v8/__init__.py
@@ -6,4 +6,4 @@ ROOT = Path(__file__).parents[0]  # yolov8 ROOT
 
 __all__ = ["classify", "segment", "detect"]
 
-from ultralytics.yolo.utils.configs import hydra_patch  # noqa (patch hydra cli)
+from ultralytics.yolo.configs import hydra_patch  # noqa (patch hydra cli)
